# artifacts/api-server/src/routes/comments.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, commentsTable, activityTable } from "@workspace/db";
import {
  ListCommentsParams,
  ListCommentsResponse,
  CreateCommentParams,
  CreateCommentBody,
  DeleteCommentParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/tasks/:taskId/comments", async (req, res): Promise<void> => {
  const params = ListCommentsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const comments = await db.select().from(commentsTable).where(eq(commentsTable.taskId, params.data.taskId)).orderBy(commentsTable.createdAt);
  res.json(ListCommentsResponse.parse(comments));
});

router.post("/tasks/:taskId/comments", async (req, res): Promise<void> => {
  const params = CreateCommentParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = CreateCommentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [comment] = await db.insert(commentsTable).values({ ...parsed.data, taskId: params.data.taskId }).returning();
  await db.insert(activityTable).values({
    type: "comment_added",
    description: `Comment added`,
    entityName: parsed.data.authorName ?? "Team member",
  });
  res.status(201).json(comment);
});

router.delete("/comments/:id", async (req, res): Promise<void> => {
  const params = DeleteCommentParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [comment] = await db.delete(commentsTable).where(eq(commentsTable.id, params.data.id)).returning();
  if (!comment) {
    res.status(404).json({ error: "Comment not found" });
    return;
  }
  res.sendStatus(204);
});

export default router;