# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NOSMO Nexus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  entityName: string;
  createdAt: Date;
}