# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NOSMO Nexus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  projectId?: number | null;
}