# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NOSMO Nexus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProjects: number;
  totalPlans: number;
  openTasks: number;
  inProgressTasks: number;
  completedTasks: number;
}