# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NOSMO Nexus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlanStatus } from './planStatus';

export interface Plan {
  id: number;
  projectId: number;
  filename: string;
  originalName: string;
  /** @nullable */
  fileSize?: number | null;
  status: PlanStatus;
  /** @nullable */
  analysisResult?: string | null;
  createdAt: Date;
}